Move endpoint for fetching location names to plan resource
