+++ src/test/java/org/opensrp/web/rest/PlanResourceTest.java
@@ -2,28 +2,40 @@ package org.opensrp.web.rest;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.opensrp.common.AllConstants;
+import org.opensrp.domain.LocationDetail;
 import org.opensrp.domain.PlanDefinition;
 import org.opensrp.domain.postgres.Jurisdiction;
+import org.opensrp.service.PhysicalLocationService;
 import org.opensrp.service.PlanService;
+import org.springframework.test.web.server.MvcResult;
 import org.springframework.test.web.server.result.MockMvcResultMatchers;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 
 import static junit.framework.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.anyList;
 import static org.mockito.Mockito.anyLong;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 import static org.opensrp.web.rest.PlanResource.OPERATIONAL_AREA_ID;
+import static org.springframework.test.web.server.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.server.result.MockMvcResultMatchers.status;
 
 /**
  * Created by Vincent Karuri on 06/05/2019
@@ -33,6 +45,8 @@ public class PlanResourceTest extends BaseResourceTest<PlanDefinition> {
     private final static String BASE_URL = "/rest/plans/";
 
     private PlanService planService;
+
+    private PhysicalLocationService locationService;
 
     private final String plansJson = "{\n" +
             "  \"identifier\": \"plan_1\",\n" +
@@ -131,8 +145,10 @@ public class PlanResourceTest extends BaseResourceTest<PlanDefinition> {
     @Before
     public void setUp() {
         planService = mock(PlanService.class);
+        locationService = mock(PhysicalLocationService.class);
         PlanResource planResource = webApplicationContext.getBean(PlanResource.class);
         planResource.setPlanService(planService);
+        planResource.setLocationService(locationService);
     }
 
     @Test
@@ -368,5 +384,22 @@ public class PlanResourceTest extends BaseResourceTest<PlanDefinition> {
 
         assertEquals(actualPlan.getIdentifier(), expectedPlan.getIdentifier());
         assertEquals(actualPlan.getJurisdiction().get(0).getCode(), expectedPlan.getJurisdiction().get(0).getCode());
+    }
+
+    @Test
+    public void testFindLocationNamesByPlanId() throws Exception {
+        LocationDetail locationDetail = new LocationDetail();
+        locationDetail.setIdentifier("304cbcd4-0850-404a-a8b1-486b02f7b84d");
+        locationDetail.setName("location one");
+
+        List<LocationDetail> locationDetails = Collections.singletonList(locationDetail);
+        when(locationService.findLocationDetailsByPlanId(anyString()))
+                .thenReturn(locationDetails);
+        MvcResult result = mockMvc
+                .perform(get(BASE_URL + "findLocationNames/{planIdentifier}", "plan_id"))
+                .andExpect(status().isOk()).andReturn();
+        verify(locationService).findLocationDetailsByPlanId(anyString());
+        Assert.assertEquals(LocationResource.gson.toJson(locationDetails), result.getResponse().getContentAsString());
+
     }
 }
